+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 ### 2020-05-xx Version 11.1.1
 
 -   Added validation for the lib configuration
-    -   Fixes [#725](https://github.com/damienbod/angular-auth-oidc-client/issues/725)
+    -   [PR](https://github.com/damienbod/angular-auth-oidc-client/pull/731) // Fixes [#725](https://github.com/damienbod/angular-auth-oidc-client/issues/725)
 -   fixed some doc typos
 -   fixed bug 2 auth events emitter on STS callback
-    -   Fixes [#734](https://github.com/damienbod/angular-auth-oidc-client/issues/734)
+    -   Fixes [PR](https://github.com/damienbod/angular-auth-oidc-client/pull/731) // [#734](https://github.com/damienbod/angular-auth-oidc-client/issues/734)
 
 ### 2020-05-14 Version 11.1.0